Removendo campo aluno_numero e testes de integração de listar

+++ src/SME.SGP.Infra/Dtos/EncaminhamentoNAAPA/FiltroEncaminhamentoNAAPADto.cs
@@ -8,7 +8,7 @@ namespace SME.SGP.Infra
         public int AnoLetivo { get; set; }
         public long DreId { get; set; }
         public string CodigoUe { get; set; }
-        public int TurmaId { get; set; }
+        public long TurmaId { get; set; }
         public string NomeAluno { get; set; }
         public DateTime? DataAberturaQueixaInicio { get; set; }
         public DateTime? DataAberturaQueixaFim { get; set; }
